+++ packages/plugins/AddressSearch/src/index.ts
@@ -8,6 +8,12 @@ import { makeStoreModule } from './store'
 import SearchResultSymbols from './utils/searchResultSymbols'
 export { SearchResultSymbols }
 
+export type {
+  AddressSearchState,
+  AddressSearchGetters,
+  MpApiParameters,
+} from './types'
+
 export default (options: AddressSearchConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'addressSearch',