+++ packages/plugins/Draw/src/store/createInteractions/createModifyInteractions.ts
@@ -4,10 +4,10 @@ import { PolarActionContext } from '@polar/lib-custom-types'
 import { CreateInteractionsPayload, DrawGetters, DrawState } from '../../types'
 
 export default function (
-  { commit, getters }: PolarActionContext<DrawState, DrawGetters>,
+  { commit, dispatch, getters }: PolarActionContext<DrawState, DrawGetters>,
   { drawSource, drawLayer }: CreateInteractionsPayload
 ): Interaction[] {
-  const { drawMode, fontSizes } = getters
+  const { drawMode } = getters
   // clear input - no feature selected initially
   commit('setTextInput', '')
   const select = new Select({
@@ -19,16 +19,14 @@ export default function (
   select.on('select', (event) => {
     if (event.selected.length > 0) {
       lastSelectedFeature = event.selected[event.selected.length - 1]
+      commit('setSelectedFeature', lastSelectedFeature)
       const featureStyle = lastSelectedFeature.getStyle()
-      if (featureStyle && 'getText' in featureStyle) {
-        const featureText = featureStyle.getText().getText()
-        const font = featureStyle.getText().getFont()
-        // set selectedSize of feature to prevent unintentional size change
-        const fontSize = font.match(/\b\d+(?:.\d+)?/)
-        commit('setSelectedSize', fontSizes.indexOf(Number(fontSize)))
-        commit('setTextInput', featureText)
-        commit('setSelectedFeature', lastSelectedFeature)
-      }
+      dispatch(
+        featureStyle && 'getText' in featureStyle && featureStyle.getText()
+          ? 'modifyTextStyle'
+          : 'modifyDrawStyle',
+        featureStyle
+      )
     } else if (event.selected.length === 0) {
       if (lastSelectedFeature && lastSelectedFeature.get('text') === '') {
         drawSource.removeFeature(lastSelectedFeature)