+++ packages/plugins/Draw/src/store/createInteractions/index.ts
@@ -1,7 +1,7 @@
-import { Draw, Snap } from 'ol/interaction'
 import createStyle from '@masterportal/masterportalapi/src/vectorStyle/createStyle'
 import Interaction from 'ol/interaction/Interaction'
 import { PolarActionContext } from '@polar/lib-custom-types'
+import { Draw, Snap } from 'ol/interaction'
 import { CreateInteractionsPayload, DrawGetters, DrawState } from '../../types'
 import createDeleteInteractions from './createDeleteInteractions'
 import createTextInteractions from './createTextInteractions'
@@ -21,8 +21,14 @@ console.warn = function (message: string) {
 export default function (
   {
     dispatch,
-    getters: { drawMode, mode, textInput, textSize },
-    rootGetters: { configuration },
+    getters: {
+      configuration,
+      drawMode,
+      mode,
+      textInput,
+      textSize,
+      selectedStrokeColor,
+    },
   }: PolarActionContext<DrawState, DrawGetters>,
   { drawSource, drawLayer }: CreateInteractionsPayload
 ): Interaction[] | Promise<Interaction[]> {
@@ -32,14 +38,19 @@ export default function (
         textInput,
         textSize,
         drawSource,
-        configuration.draw
+        configuration
       )
     }
-    const style = configuration.draw
-      ? configuration.draw[
+    const style = configuration
+      ? configuration[
           `${drawMode.charAt(0).toLowerCase()}${drawMode.slice(1)}Style`
         ]
       : undefined
+    // TODO const style = createDrawStyle(
+    //   drawMode,
+    //   selectedStrokeColor,
+    //   configuration?.style
+    // )
     const draw = new Draw({ source: drawSource, type: drawMode })
     const vectorStyle = { rules: [{ style }] }
     if (style) {