+++ packages/clients/meldemichel/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Fix: Import type `MpApiParameters` from correct position.
 - Chore: Change value of `pins.movable` configuration to `'drag'` as using a boolean has been deprecated in a future release.
 
 ## 1.1.2